requirements.txt
python-telegram-bot==20.7
SQLAlchemy
aiosqlite
python-dotenv